backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-mongoengine==1.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
